+++ packages/ui/src/utils/JSXTypeUtil.ts
@@ -24,6 +24,7 @@ import type { WuiCtaButton } from '../composites/wui-cta-button/index.js'
 import type { WuiDetailsGroup } from '../composites/wui-details-group/index.js'
 import type { WuiDetailsGroupItem } from '../composites/wui-details-group-item/index.js'
 import type { WuiEmailInput } from '../composites/wui-email-input/index.js'
+import type { WuiEnsInput } from '../composites/wui-ens-input/index.js'
 import type { WuiIconBox } from '../composites/wui-icon-box/index.js'
 import type { WuiIconLink } from '../composites/wui-icon-link/index.js'
 import type { WuiInputAmount } from '../composites/wui-input-amount/index.js'
@@ -104,6 +105,7 @@ declare global {
       'wui-details-group-item': CustomElement<WuiDetailsGroupItem>
       'wui-details-group': CustomElement<WuiDetailsGroup>
       'wui-email-input': CustomElement<WuiEmailInput>
+      'wui-ens-input': CustomElement<WuiEnsInput>
       'wui-icon-box': CustomElement<WuiIconBox>
       'wui-icon-link': CustomElement<WuiIconLink>
       'wui-input-amount': CustomElement<WuiInputAmount>